pasar la portada del video a component
